Refactor Steam API calls to SteamWebApiHelper
Moved the makeApiCall logic from SocialServiceHelper to a new SteamWebApiHelper class for better separation of concerns and reusability. Updated SocialServiceHelper and SteamSocialService to use SteamWebApiHelper for all Steam Web API requests.
